LayoutTests/inspector-obsolete/console/console-object-preview-expected.txt

CONSOLE MESSAGE: line 9: Mutating object in a loop
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 13: [object Object]
CONSOLE MESSAGE: line 16: Mutating array in a loop
CONSOLE MESSAGE: line 20: 0,0,0
CONSOLE MESSAGE: line 20: 0,0,1
CONSOLE MESSAGE: line 20: 0,0,2
CONSOLE MESSAGE: line 23: Object with many properties
CONSOLE MESSAGE: line 28: [object Object]
CONSOLE MESSAGE: line 30: Array with many properties
CONSOLE MESSAGE: line 35: 0,1
CONSOLE MESSAGE: line 37: Object with proto
CONSOLE MESSAGE: line 40: [object Object]
CONSOLE MESSAGE: line 42: Sparse array
CONSOLE MESSAGE: line 45: ,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,50,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
CONSOLE MESSAGE: line 47: Dense array with indexes and propeties
CONSOLE MESSAGE: line 53: 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143,144,145,146,147,148,149
Tests that console produces instant previews for arrays and objects.

Mutating object in a loop console-object-preview.html:9 console-message-text source-code > console-message-url webkit-html-resource-link
Object {a: 0, b: 0, c: 0} console-object-preview.html:13 console-message-text source-code > console-formatted-object source-code > console-object-preview > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > console-message-url webkit-html-resource-link
Object {a: 0, b: 0, c: 1} console-object-preview.html:13 console-message-text source-code > console-formatted-object source-code > console-object-preview > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > console-message-url webkit-html-resource-link
Object {a: 0, b: 0, c: 2} console-object-preview.html:13 console-message-text source-code > console-formatted-object source-code > console-object-preview > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > console-message-url webkit-html-resource-link
Mutating array in a loop console-object-preview.html:16 console-message-text source-code > console-message-url webkit-html-resource-link
[0, 0, 0] console-object-preview.html:20 console-message-text source-code > console-formatted-array source-code > console-object-preview > console-formatted-number > console-formatted-number > console-formatted-number > console-message-url webkit-html-resource-link
[0, 0, 1] console-object-preview.html:20 console-message-text source-code > console-formatted-array source-code > console-object-preview > console-formatted-number > console-formatted-number > console-formatted-number > console-message-url webkit-html-resource-link
[0, 0, 2] console-object-preview.html:20 console-message-text source-code > console-formatted-array source-code > console-object-preview > console-formatted-number > console-formatted-number > console-formatted-number > console-message-url webkit-html-resource-link
Object with many properties console-object-preview.html:23 console-message-text source-code > console-message-url webkit-html-resource-link
Object {property_0: 0, property_1: 1, property_2: 2, property_3: 3, property_4: 4…} console-object-preview.html:28 console-message-text source-code > console-formatted-object source-code > section > header monospace > subtitle > title > console-object-preview > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > object-info-state-note > properties properties-tree monospace > console-message-url webkit-html-resource-link
Array with many properties console-object-preview.html:30 console-message-text source-code > console-message-url webkit-html-resource-link
[0, 1, property_0: 0, property_1: 1, property_2: 2, property_3: 3, property_4: 4…] console-object-preview.html:35 console-message-text source-code > console-formatted-array source-code > section > header monospace > subtitle > title > console-object-preview > console-formatted-number > console-formatted-number > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > object-info-state-note > properties properties-tree monospace > console-message-url webkit-html-resource-link
Object with proto console-object-preview.html:37 console-message-text source-code > console-message-url webkit-html-resource-link
Object {d: 1, a: 0, b: 0, c: 2} console-object-preview.html:40 console-message-text source-code > console-formatted-object source-code > console-object-preview > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > name > console-formatted-number > console-message-url webkit-html-resource-link
Sparse array console-object-preview.html:42 console-message-text source-code > console-message-url webkit-html-resource-link
[50: 50] console-object-preview.html:45 console-message-text source-code > console-formatted-array source-code > console-object-preview > name > console-formatted-number > console-message-url webkit-html-resource-link
Dense array with indexes and propeties console-object-preview.html:47 console-message-text source-code > console-message-url webkit-html-resource-link
[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99…] console-object-preview.html:53 console-message-text source-code > console-formatted-array source-code > section > header monospace > subtitle > title > console-object-preview > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > console-formatted-number > object-info-state-note > properties properties-tree monospace > console-message-url webkit-html-resource-link
